[MediaPrivacy] Make is_sensitive value respect the flag
The is_sensitive field in MediaSession is used as a proxy to hide
sensitive metadata from more public surfaces such as the lock screen.
Since we are replacing the sensitive metadata itself with placeholders,
we no longer will need to rely on this field to completely hide the
metadata or the media controls.

Fixed: b:301593406
Bug: b:290201426, 1447545
Change-Id: Ic0fc993ff29108c96ada1549d65a36925d407d73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4887238
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Filipa Senra <fsenra@google.com>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203150}