[ios] Move key ActivityBucketLastReportedDate out of NSUserDefaults
This CL moves kActivityBucketLastReportedDateKey out of NSUserDefaults.

The key has been moved in the PrefService. Note that there are two
types of PrefServices, in this case the code has been moved in the
PrefService coming from BrowserState which is per account and is
syncable between devices.

Bug: 1486770
Change-Id: I2308e931e507958b9f1cea9a4fa7dc0011666ca8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4803825
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1203163}