android: Add ColdStartTracker, use in CustomTabObserver
Add new class ColdStartTracker utility for telling whether the first
Activity was created at cold start. Use this utility in
CustomTabObserver to record new CCT metrics:

    CustomTabs.Startup.TimeTo*

Here we extract the cold start detection logic from these metrics, and
put them to the ColdStartTracker. In addition we extend it with a
time-based heuristic for devices running N-O. The metrics were not
recorded on this platform, but still we prefer to abstract away from
this fact in the code, so we enable collecting the metrics on N-O and
document the related caveats in the metric descriptions.

The plan is to use the ColdStartTracker in other places. For example, it
can be used to decide whether Paint Preview can be shown (it is
currently shown only on cold starts, using a different heuristic).

Bug: 1476038
Change-Id: Ic1a29f7431b8dc8809abcc0fb796bdb54a6c2de8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4887183
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Ella Ge <eirage@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203165}