Move touchscreen_tap_suppression_controller to //content/common/input
This refactoring is a part of bigger project to handle input on viz
thread (go/input-on-viz). We are moving input related interfaces that
RenderWidgetHostImpl uses to //content/common/input. This would provide
a single stack in //content/common/input that we quickly enter in either
the browser or GPU viz which would allow for relatively clean movement
of code. GPU process can link in the same input stack and forward the
data to the renderer using the same code.

Bug: b/294828208
Change-Id: Id079b776ce8fe8796af27fc9620bb7d387db3167
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4774207
Commit-Queue: Aman Verma <amanvr@google.com>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203187}