[mullet m2] Open EditAddressProfileView via controller.
Previously, EditAddressProfileView was opened from
AutofillBubbleHandler. The opening logic doesn't depend on the
ToolbarButtonProvider inside the AutofillBubbleHandler. Thus this
dialog should not be opened through the AutofillBubbleHandler in the
first place.

This CL moves to opening logic inside the
EditAddressProfileDialogController. Unit tests need a way to override
this logic so that the view is not created. This CL introduces a view
factory, which is set to override the default behavior.

To avoid referencing the c/b/ui/views concepts from the c/b/ui, a
special header file is created called
c/b/ui/autofill/edit_address_profile_view.h. It declares a method to
open the view. The implementation is placed inside the
c/b/ui/views/autofill/edit_address_profile_view.cc

Bug: 1459990
Change-Id: Ie916a48d8a333f5f2ad19e27ad7c0635c5b932fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4905244
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1203198}