[UPMLocalSettings] Remove the old settings migration pref
The UPM migration for the syncing users is done.
The migration pref was being written to, but not used for re-enrollment,
so it's no longer useful.

OBSOLETE_HISTOGRAMS[PasswordManager.MigratedSettingsUPMAndroid]= The
histogram is no longer needed because the UPM settings migration
strategy has changed.

Bug: 1486451
Change-Id: I17954d5794de2f5f42c17059d90b9f12993c6972
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4887476
Commit-Queue: Ivana Žužić <izuzic@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203212}