[IP Protection] Get Anonymous Tokens library building
The Anonymous Tokens library is moving towards no longer being
released alongside Quiche, so as the first step towards that,
get the AT library building and the tests passing separately.

cl/567348596 has the remaining changes to make this split in
Quiche on Github, and then in the next roll we'll incorporate
the changes needed to complete this on the Chrome side (see:
https://chromium-review.googlesource.com/c/chromium/src/+/4549362/5/net/third_party/quiche/BUILD.gn)

I'm still working through an issue with getting some of the tests
passing with the ios-simulator trybot. The issue is likely related
to the testdata path or the corresponding bundle_data, but what's
weird is that I can't reproduce these failures locally (all tests
pass) and the corresponding GN config is very similar to what
Quiche was using (where the tests passed successfully). I'll
investigate this in a follow-up CL so that it doesn't block this
one.

Bug: 1446453
Change-Id: I901fb53d7526a2d5b8c917c1a01da6f4c2d66af0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4885470
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203218}