Add VariableRefreshRateController
This CL creates the VariableRefreshRateController to manage the enabled
or disabled state of the variable refresh rate feature. VRRC observes
the GameMode state, BatterySaverMode state, and relevant feature flags
for determining whether VRR should be enabled. Updates are issued to
the DisplayConfigurator for initiating modesets as needed.

This CL also adds an Observer interface to GameModeController so that
state changes can be subscribed to by VRRC and others.

Bug: b:298427287, b:297345871
Change-Id: I1dfdda8967d9e09d5930506cf3c7edee03d41a84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4886675
Reviewed-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: Andrew Wolfers <aswolfers@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203228}