IndexedDB: remove IndexedDBDispatcherHost
The last responsibilities, which are to implement the
blink::mojom::IDBFactory interface, are moved to IndexedDBFactory.
This is not only a more intuitive place for the functionality, it's
also where the meat of the functionality was already implemented.

The majority of the complexity of this change is in updating/upgrading
tests to use the public Mojo API instead of internal APIs. This should
make affected tests less brittle to future changes and the system will
be easier to iterate. In particular, using Mojo APIs and testing
framework makes async testing more natural, and future updates to
synchronocity, e.g. due to sharding backing stores across threads,
will not disrupt tests that assume synchronous results.

Bug: 1474996

Change-Id: I07901a400a0653fc02dbd8c469d053e6531d347f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4888631
Reviewed-by: Austin Sullivan <asully@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203264}