[ios] Change SessionRestorationObserver API
In order to allow using SessionRestorationObserver with both the
legacy and the optimised session restoration code, change the API
to take the Browser as an additional parameter.

Keep the old API as it is still used in some internal code. The
public observer will be updated in a followup CL once the internal
code has been fixed.

Add method to register SessionRestorationObserver with the new
SessionRestorationService (for use when the optimised session
restoration feature is enabled).

Add helper functions to register a SessionRestorationObserver
with the correct object depending on which version of the session
restoration code is being used (based on the feature flag).

This is part of a multi-CLs change:
  => https://crrev.com/c/4905372
  -  https://crrev.com/i/6506796
  -  https://crrev.com/c/4902850

Bug: 1383087
Change-Id: I9ee34bdfdaaefbe9c4823eabe00cefe4d8f432fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4905372
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1203289}