[//gpu] Add glTexImage2DSharedImageCHROMIUM() GLES2 extension
This CL adds the following API to the GLES2 command decoders:

GLuint glTexImage2DSharedImageCHROMIUM (GLunit texture,
                                        const GLbyte *mailbox)

`texture` is a client-side texture, and `mailbox` is the SharedImage to
assign to the first level of the texture.

This function maps `texture` to the texture object of the SharedImage
represented by the mailbox and ensures that if `texture` is bound to a
target, the service-side texture gets bound to that target.

On success, this function extends the lifetime of the SharedImage
referenced by mailbox by taking a ref. The texture object is deleted and
the ref released once the calling context deletes the texture name
associated with the texture object, and detaches it from all framebuffer
objects as well as texture unit bindings.

This CL also implements the new API in the passthrough decoder. The
implementation follows that of
glCreateAndTexStorage2DSharedImageCHROMIUM(), with the differences
being that

(a) the latter creates and returns a new client texture ID whereas the
    new function uses the client texture ID passed in from the client
(b) TexImage2DSharedImageCHROMIUM() binds the client texture ID to the
    target, as it might have previously been bound to a different
    target.

Implementation in the validating decoder will happen as a followup.

This API will shortly be used to implement GPU-GPU copy of
VideoFrame textures to a WebGL texture via a temporary SharedImage that
is attached to that WebGL texture via this API.

Bug: 1410164
Change-Id: I875ada9abec08074f50f9709c0546b7a5303e73f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4900518
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203293}