[Tab Switcher] Fix lazy creation on tablets and add test
When Start Surface Refactor is enabled and Defer Tab Switcher Layout
Creation is enabled there is a bug where Tab Switcher creation will
fail to lazily be created for Tablet because the codepath for Tablets
re-uses createTabSwitcherOrStartSurface. On phones createGridTabSwitcher
is called directly, but we can't align these behaviors until
Start Surface Refactor is launched and Tablet no longer needs to create
a Start Surface Layout.

This issue was introduced in:
https://chromium-review.googlesource.com/c/chromium/src/+/4873674

I should have manually tested this and I thought there was automated
coverage that wasn't showing up. However, because the feature is behind
a cached feature flag it doesn't trigger regardless of whether it is
enabled by default in the fieldtrial testing config.

I've added an explicit test with the flag enabled that should catch
the regression and manually tested to confirm the crash is resolved.

Cq-Include-Trybots: luci.chromium.try:android-12l-x64-dbg
Fixed: 1487114
Change-Id: Iff78eace7cbddeb3dbc9e35c8b2ccdf7211f684d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4897299
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Cr-Commit-Position: refs/heads/main@{#1203297}