[iOS] Prevent external intents if the SignedIn screen is displayed
When the accounts on the devices are updated, Chrome is displaying a
screen letting the user know of the new accounts currently available.
When this screen is presented, the intents should not be handled as
they might not interact well with the feature.

Bug: 1476453
Change-Id: Ida36b21b1753442c5122c68f07934cf10d6ca6fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4903786
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203346}