[FedCM] Embargo permissions upon closing mismatch dialog
If the mismatch UI is closed by clicking the "x" button or swiping down
to close, we would embargo permissions for third-party sign-in for the
RP origin. In other cases such as clicking on the continue button in the
mismatch UI then closing the pop-up window, we would not embargo the
origin.

Bug: 1454803
Change-Id: I4fb1657fdf785c97359606953cf93b4e6edf77f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4902608
Commit-Queue: Zachary Tan <tanzachary@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203366}