[Nearby][Self Share] Fix NS browser tests to account for Self Share
All tests in the `browser_tests`, `unit_tests`, and `ash_unittests`
targets were run with the Self Share flag enabled. This CL fixes the
tests that failed with Self Share was enabled and not disabled.
The reast of the failures are chalked up to errors in ToT.

When these pass, the finch config to roll out Self Share can merge:
crrev/c/4898928

Bug: b/302024934
Change-Id: I0dfe9fc01767b2eec2d7359a9bb4f26e3183d6a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4902970
Reviewed-by: Daniel Classon <dclasson@google.com>
Commit-Queue: Brando Socarras <brandosocarras@google.com>
Cr-Commit-Position: refs/heads/main@{#1203391}