FSA: Rename locking scheme and add base::Feature
Removes "AccessHandle" from the fsa locking scheme flag since it does
not only apply to access handles. Now generates a base::Feature. Added
a few CHECKs on the browser side to assert the disabled flag behavior.

Also removes the experimental feature status since this causes issues
with the base::Feature and RuntimeEnabledFeature not being equivalent
in some situations.

Adds a virtual test suite to enable this flag for locking scheme WPTs.
The non virtual test version of them were added to the
TestExpectations to fail rather than setting them to run exclusively
in virtual tests. This is because of a bug with the linting of the
VirtualTestSuites file (crbug.com/1399685).


Bug: 1382215
Change-Id: I16d400db8d922cea2154fe2edcff02414b30eb69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4895799
Reviewed-by: Austin Sullivan <asully@chromium.org>
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203401}