[Android] Inline methods from CachedFeatureFlags
CachedFeatureFlags's logic is being distributed to other classes. It
is meant to be deleted soon.

Also move the header comment of CachedFeatureFlags since now
it should not be referenced very often, and move the header to
CachedFlag, now the main class for CachedFlags.

Bug: 1442347,1487364
Change-Id: Id1c20a948967ceb3164a38c74dd0eef286c678f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4860422
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203404}