Correct missing PrintBackendServiceManager client registrations
Move the check for registered clients in GetService() to be the first
thing this method does.  This allows tests to expose other paths that
are otherwise missing registrations with PrintBackendServiceManager.
Fill those in to correct usage with PrintBackendServiceManager.

Bug: b/301590015
Change-Id: I7401141bb41c0abd734470564b05a44d2167cbb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4898532
Reviewed-by: Andrew Rayskiy <greengrape@google.com>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203405}