focus-mode: Add focus panel timer setting UI
This CL adds a section to the focus panel that allows for the user to
set the timer when they are not in a focus session. It has a textbox for
setting exact amounts of time and increment and decrement buttons for
adjusting the time. We also add unit tests to verify that it works
correctly.

Bug: b/297077897
Change-Id: If5bc7bd51e7d468f460a38cb7377c4b0187251a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4885347
Commit-Queue: Ben Becker <benbecker@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203437}