Send b64-encoded compressed safe seed to Platform on successful fetch
Currently, Chrome is sending the serialized but not encoded nor
compressed form of the safe seed to Platform on successful fetch. This
was implemented in https://crrev.com/c/4261237.

Instead, send the b64-encoded compressed form of the safe seed so
evaluate_seed logic can more closely resemble the current Chrome browser
flow and also save disk space when storing the seed.

Most of the code in this CL previously existed in SafeSeedManager and is
moved to VariationsSeedStore so the b64-encoded compressed safe seed can
be accessed and sent to the Platform.

One new requirement added in this CL is that the safe seed will only be
sent to Platform if the seed is correctly validated.

BUG=b:300996158
TEST=included unit tests

Change-Id: Iee5aa0f9e2d63df23af357fcbd5131654e268f6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4877853
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Commit-Queue: Kendrake Tsui <kendraketsui@google.com>
Cr-Commit-Position: refs/heads/main@{#1203443}