Avoid undefined behavior in AudioFifo
media::AudioFifo maintains two signed integer counters for calculating
the current number of frames in the buffer: frames_pushed_ and
frames_consumed_.  These are incremented but never decremented, so
after many pushes and pulls they will overflow.  Signed integer
overflow is undefined in C++.  While this likely isn't causing any
actual issues on the platforms we support, we should probably avoid
undefined behavior when possible.

We could fix this by converting the counters to unsigned, but we can
also use one counter and increment or decrement it as needed. This
class is already marked as thread-unsafe so we don't have to consider
concurrency issues, and since the maximum number of frames is also
stored in an int this counter will never overflow.

Change-Id: Ib8aec86180214d5144b78fc416e943448d2d49a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4903290
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203479}