[EditContext] Dispatch textformatupdate when composition ends
The textformatupdate event is fired during composition to indicate to
the developer that they need to apply styles to the text per request
of the IME. But, we fail to fire the event again to clear the styles
when the composition is ended.

Developers could work around this by clearing the formats on the
compositionend event, but for consistency's sake we should fire
textformatupdate to indicate to the developer that the formats
should be cleared.

Bug: 999184
Change-Id: Ia3ff56b3376718820fd092892a3ca440cd6d2332
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4895644
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1203507}