Provide a way to assign an identifier to a menu.
This will be useful for testing; especially for dynamically-populated
menus where a specific element with a specific identifier isn't
guaranteed to be present.

This is not a complete solution; there may be other points at which we
would want to inject an identifier, but it does create a general pattern
by which such identifiers could be assigned (e.g. by adding to some of
the existing menu model and adapter classes in ui/ and views/).

Implementation notes:

The challenge with injecting an identifier here is that the submenu is
created by its parent menu item, not the model adapter/runner. Also,
the root submenu is created by a menu item that is never actually added
to a widget; it serves only as a placeholder.

Therefore, the easiest way to convey the identifier to the submenu is
via the parent MenuItemView.

TODO(dfried): need to add Mac support for native context menus. Fix before submitting.

Change-Id: I2b223f4befa257b0c8244eff446463750f6b0160
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4893108
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203522}