Attach appropriate SBS to Horizontally rendered suggestions.
The change alters how Searchbox Stats are generated when a
Horizontall render group is in use.

Currently, the logic assumes are AutocompleteMatch objects are
arranged vertically in a linear fashion. This creates a direct, 1:1
correspondence between AutocompleteMatch index in the
AutocompleteResult object, and its position reported via
SearchboxStats.

With Horizontal render groups, the relationship changes. Multiple
AutocompleteMatch objects may reside at the same horizontal position.

In order to report appropriate SBS for each of the horizontally
rendered suggestions, the change allocates a single SuggestionInfo
slot and updates its content for every of the AutocompleteMatches that
target this specific slot.

Bug: 1474087
Change-Id: If8dc5f71e71e440e344ba07df1be2a93578b9033
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4905794
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1203531}