[Extensions] Ensure ImageWriterControllerLacros API factory is built.
The factory for this BrowserContextKeyedAPI instance needs to be built
during Profile initialization. The step was omitted when the class was
introduced.

This issue wasn't detected when this class was implemented because
there are no tests for the imageWriterPrivate API on Lacros. I
briefly investigated enabling the test for Lacros, but it doesn't
seem straightforward.

To make sure this CL fixes the issue, I did the following:

1. Added the API test file to the build for Lacros.
2. Modified one of the API tests to retrieve the
ImageWriterControllerLacros instance and assert it's not a nullptr.
3. Ran the test and observed that the DCHECK in DependencyManager
that verifies that the factory was built was triggered.
4. Implemented the fix.
5. Rebuilt and ran the test again.
6. Verified that the instance is returned as expected.

Bug: 1483219
Change-Id: If389c5bfa14e5d78d27f81bd480212113044b208
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4884759
Reviewed-by: Tim <tjudkins@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203533}