[GridNG] Return shortest possible grid-lines with <custom-ident>'s
In order to return the shortest-possible representation for `grid-rows`,
`grid-columns`, and `grid-area`, we need to omit duplicate
<custom-ident> values in certain shorthand scenarios. This is
unfortunately more complex than just dropping subsequent values, as
they are not always adjacent per spec.

The simplest option here was to drop the shared loop for
shorthands in this case, in favor of specialized methods for `grid-area`
and `grid-column/row` that enforces the order of fallback behavior
outlined in the spec.

cssom_utils.h/cc was added for sharing code.

I updated the fast/css-grid-layout expectations, and now they are
all 100% interoperable with Firefox, finally!

Bug: 1484952, 1484855
Change-Id: I58193f52e66f9a4cc2c372ee47955fcc30999125
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4894897
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1203546}